requirements.txt
Django==2.2
django-crispy-forms==1.9.0
django-phone-field==1.8.0
Pillow==7.1.2
pytz==2020.1
sqlparse==0.3.1
